Adds a test case for api.py. The common fixture is added to tests/fixtures
